[CQ] Numeric issues: fix implicit casts (#8005)
Implicit casts can lead to lossy conversions.

+++ flutter-idea/src/io/flutter/view/MultiIconSimpleColoredComponent.java
@@ -423,7 +423,7 @@ public class MultiIconSimpleColoredComponent extends JComponent implements Acces
 
     final Font font = getBaseFont();
 
-    width += computeTextWidth(font, mainTextOnly);
+    width += (int)computeTextWidth(font, mainTextOnly);
     width += myIpad.right + borderInsets.right;
 
     // Take into account that the component itself can have a border
